Update the model for Postgresql support

+++ db/regtests/mysql/create-ado-mysql.sql
@@ -32,7 +32,7 @@ CREATE TABLE TEST_COMMENTS (
   /* the comment identifier */
   `ID` INTEGER ,
   /* the comment version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* the comment publication date. */
   `DATE` TIMESTAMP NOT NULL,
   /* the comment message. */
@@ -50,7 +50,7 @@ CREATE TABLE test_image (
   /* the image identifier */
   `id` INTEGER ,
   /* the image version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* the message creation date */
   `create_date` DATETIME NOT NULL,
   /* the image data */
@@ -62,7 +62,7 @@ CREATE TABLE allocate (
   /* the user id */
   `ID` BIGINT NOT NULL,
   /* the allocate version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* the sequence value */
   `NAME` VARCHAR(255) ,
   PRIMARY KEY (`ID`)
@@ -72,7 +72,7 @@ CREATE TABLE test_user (
   /* the user id */
   `ID` BIGINT NOT NULL,
   /* the comment version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* the sequence value */
   `VALUE` BIGINT NOT NULL,
   /* the user name */
@@ -86,7 +86,7 @@ CREATE TABLE test_nullable_table (
   /* the user id */
   `ID` BIGINT NOT NULL,
   /* the comment version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* an identifier value */
   `ID_VALUE` BIGINT ,
   /* an integer value */
@@ -106,13 +106,13 @@ CREATE TABLE test_table (
   /* the user id */
   `ID` BIGINT NOT NULL,
   /* the comment version. */
-  `version` int NOT NULL,
+  `version` INTEGER NOT NULL,
   /* an identifier value */
   `ID_VALUE` BIGINT NOT NULL,
   /* an integer value */
   `INT_VALUE` INTEGER NOT NULL,
   /* a boolean value */
-  `BOOL_VALUE` INTEGER NOT NULL,
+  `BOOL_VALUE` TINYINT NOT NULL,
   /* a string value */
   `STRING_VALUE` VARCHAR(255) NOT NULL,
   /* a time value */
